+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Multitenant AWS Dynamo supports the [AWS Dynamo Java API](http://docs.aws.amazon
   
 You can write your application code against the Amazon DynamoDB interface as you would for any other application.  The implementation will manage storage of data by tenant.
 
+## 0.9.31 (March 25, 2019)
+* Fixed bug in `CachingAmazonDynamoDbStreams` where an `IllegalArgumentException` was thrown with "Invalid position 
+segment in shard iterator string" due to an invalid iterator format
+
 ## 0.9.30 (March 14, 2019)
 
 * Added support for Billing Mode (Pay Per Request or default Provisioned)